requirements.txt
arrow==1.3.0
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-import-export==3.3.8
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.1
et-xmlfile==1.1.0
itsdangerous==2.2.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.3
rq==1.15.1
rq-win @ git+https://github.com/michaelbrooks/rq-win.git@c6e65e38f8a6ca99a0c8e0dd6fe35f9b92837cca
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
times==0.7
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
xlrd==2.0.1
xlwt==1.3.0